Convert dependency management to Poetry
Poetry has a number of advantages over Pipfile, predominantly that it
also removes the need for other config files like setup.py, MANIFEST.in,
setup.cfg, etc. Arguably it also has a nicer CLI experience that Pipenv.
For Python projects which are intended for release onto PyPi in
particular it's particularly useful.