Start moving towards an approach with first class objects
Previously the client mainly printed out serialized objects to stdin. It
would be preferable as a client to return objects, which can be
understood by consumers and manipulated more directly. This commit
changes a few of the methods and introduces a couple of classes.

In doing this I've also introduced basic error handling into the request
helper methods, and moved some of the repeated effort up into those
helpers.