Make the raw HTTP methods part of the public interface
The base methods are really useful for dropping down to talk with the
API directly, or for support of parts of the API that don't yet have a
higher-level object or method. I think that warrants them being
considered a first class part of the API rather than as an
implementation detail.