Add helper methods to get individual instances of objects
It's unclear whether the API has proper resourceful endpoints for
individual resources. And I'm on a plane so can't easily check. For the
moment that means I'm looping over the full lists and extracting
individual items. Hopefully the implementation can change but the
interface should be correct. Could also be extended to the other
objects.