Stub out unsupported bits of the API
The client doesn't yet have support for the full API. This commit adds a
few likely methods and an exception to track adding support for those.
This isn't all of the rest of the API but it's most of it. Good idea of
the scope of the full API and what's missing.