Fix the test methods so they return the results of the tests
Tests return IssueSets. Converted the relevant examples over to using
the returned pbjects.

The test methods live on organization, as they need an organization ID.
However, they aren't scoped to organization. This appears to be purely
for authorisation or similar. I wonder if allowing clients to be
instantiated with an organisation would make for a nicer API for these
utilities?