Add explicit manifest support to the high-level import method
Previously Snyk would detect a single manifest, which for multi-manifest
projects makes this method unsuitable. The changes support a backwards
compatible addition to the API to specify one or more manifest files.

This area of the code was also one of the last lacking tests. This
commit also adds some around the main methods, with a pattern that can
be expanded across the other import methods.