Enable using the Snyk CLI from Python
Both test and monitor are only available via the CLI today, and using
them programatically can be useful. This commit adds a new class which
exposes methods for running those, by shelling out to the Snyk CLI
installed locally.

The methods return a SnykRunResult object, rather than just the raw
JSON. This is to make it easy to add attributes to that response
in a backwards compatible manner.