Additional settings (#105)
* Update dependency packages to latest versions

* Speed up tests by removing unneeded delay

The delays in the tests added 20s or so to the test run, balooning it
from <2s to ~22s. Dropped the delay as not relevant to testing mocked
endpoints.

Also fixed the types which where throwing a warning. No need to be
optional given the concrete default values.

* Updated list of settings.

The API docs and the API had a disagreement. This _should_ be the
correct list of currenly available settings.