Update dependencies to support Python 3.10 (#130)
* fix: Remove incorrect usage of InitVar

Fields marked with InitVar should only be used in the __post_init__
method of a dataclass. Newer versions of mypy enforce this.

* chore: Update dependencies to work with python 310

Most of this was automatic updates with poetry update. A few
dependencies were manually updated as they required major version
changes. There are no known breaking changes between these versions.

* chore: Remove deprecated poetry-version dependency

poetry-version is deprecated and should not be used any more. Removing
this means we will be relying on importlib_metadata which has a minimum
python version dependency of 36.

* chore: Bump version number

Updating the dependencies was quite a big change so to be safe increase
the minor version number.