fix: Change vulns property so it is not deprecated
The vulnerabilities property on a project was using the deprecated (and
removed) issue set. This has been moved over to aggregated issues.

The data has been mapped to maintain the public api of this package
however small differences may be present.

This does mean there is an extra api call PER ISSUE which could be very
slow for some projects.