Fix: project vulnerabilities (#134)
* chore: Mark IssueSet as deprecated

The endpoint behind this property has been removed from GA so it cannot
be used. Alternative is to the AggregatedIssues instead which provides
mostly the same data.

* fix: Correct typings on IssueData

The typings here were not fully accurate, this patch makes them closer
to reality. Still a lot of Any in there which is ugly but I'm not
touching those paths...

* feat: Add manager for IssuePaths

The issue paths resource is a missing piece of information that used to
be exposed from the IssueSet resource, but that has been removed so we
need to fetch it explicitly when needed.

* fix: Change vulns property so it is not deprecated

The vulnerabilities property on a project was using the deprecated (and
removed) issue set. This has been moved over to aggregated issues.

The data has been mapped to maintain the public api of this package
however small differences may be present.

This does mean there is an extra api call PER ISSUE which could be very
slow for some projects.

* chore: Bump release version