Fix issue with limit duplication leading to errors
[this
commit](https://github.com/snyk-labs/pysnyk/commit/dc75bc926aa369d7d39e89e744b3b07d6af0fb0f)
introduced a default limit in the client. However, the limit is also
present in the URL. That leads to the 2nd and subsequent pages failing
due to passing `limit=X&limit=X`. That is interpreted as an array by the
server, and will fail with a client error.

The fix checks the URL for the limit, and avoids the duplication if
present.